[blinux-make] NODEL env var prevents deletion of mouli folder when set

+++ blinux-make/make_and_test.sh
@@ -31,4 +31,8 @@ cp mouli/note.txt .
 cp mouli/tests_report.xml .
 
 # wipe mouli folder
-rm -rf mouli
+if [ -z ${NODEL+x} ]
+then
+	rm -rf mouli
+fi
+
